move haystack indexing to better-structured celery tasks, allowing more to be indexed
